Reorganize transactionTimeSeries queries, always use host-metrics.js when querying for a specific host
